convert code style to MVC style
