+++ frameworks/Java/vertx-web/src/main/java/io/vertx/benchmark/App.java
@@ -1,524 +1,28 @@
 package io.vertx.benchmark;
 
-import com.github.susom.database.Config;
-import com.github.susom.database.ConfigFrom;
-import com.github.susom.database.DatabaseProviderVertx;
-import com.github.susom.database.DatabaseProviderVertx.Builder;
-import com.github.susom.database.SqlInsert;
-import io.vertx.benchmark.model.Fortune;
-import io.vertx.benchmark.model.Message;
-import io.vertx.benchmark.model.World;
-import io.vertx.core.*;
-import io.vertx.core.http.HttpHeaders;
-import io.vertx.core.json.Json;
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.mongo.MongoClient;
+import io.vertx.benchmark.benchmarks.*;
+import io.vertx.core.AbstractVerticle;
 import io.vertx.ext.web.Router;
-import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.templ.rocker.RockerTemplateEngine;
-import io.vertx.pgclient.PgConnectOptions;
-import io.vertx.pgclient.PgPool;
-import io.vertx.sqlclient.*;
-
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.*;
-
-import static io.vertx.benchmark.Helper.randomWorld;
 
 public class App extends AbstractVerticle {
 
-  /**
-   * MongoDB implementation
-   */
-  private final class MongoDBBenchmark {
-    private final JsonObject FIELDS = new JsonObject().put("_id", 0);
-
-    private final MongoClient database;
-    // In order to use a template we first need to create an engine
-    private final RockerTemplateEngine engine;
-
-    public MongoDBBenchmark(Vertx vertx, JsonObject config) {
-      final JsonObject mongoConfig = config.copy();
-
-      // mongo is configured without credentials
-      mongoConfig.remove("username");
-      mongoConfig.remove("password");
-
-      this.database = MongoClient.createShared(vertx, mongoConfig);
-      this.engine = RockerTemplateEngine.create();
-    }
-
-    public final void dbHandler(final RoutingContext ctx) {
-      database.findOne("world", new JsonObject().put("_id", randomWorld()), FIELDS, findOne -> {
-        if (findOne.failed()) {
-          ctx.fail(findOne.cause());
-          return;
-        }
-
-        ctx.response()
-            .putHeader(HttpHeaders.SERVER, SERVER)
-            .putHeader(HttpHeaders.DATE, date)
-            .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-            .end(Json.encodeToBuffer(new World(findOne.result())));
-      });
-    }
-
-    public final void queriesHandler(final RoutingContext ctx) {
-      final int queries = Helper.getQueries(ctx.request());
-
-      final World[] worlds = new World[queries];
-
-      new Handler<Integer>() {
-        @Override
-        public void handle(Integer idx) {
-          if (idx == queries) {
-            // stop condition
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encodeToBuffer(worlds));
-
-          } else {
-
-            final Handler<Integer> self = this;
-
-            database.findOne("world", new JsonObject().put("_id", randomWorld()), FIELDS, findOne -> {
-              if (findOne.failed()) {
-                ctx.fail(findOne.cause());
-                return;
-              }
-
-              worlds[idx] = new World(findOne.result());
-              self.handle(idx + 1);
-            });
-          }
-        }
-      }.handle(0);
-    }
-
-    public final void fortunesHandler(final RoutingContext ctx) {
-      final List<Fortune> fortunes = new LinkedList<>();
-
-      database.find("fortune", new JsonObject(), find -> {
-        if (find.failed()) {
-          ctx.fail(find.cause());
-          return;
-        }
-
-        for (JsonObject document : find.result()) {
-          fortunes.add(new Fortune(document));
-        }
-
-        fortunes.add(new Fortune(0, "Additional fortune added at request time."));
-        Collections.sort(fortunes);
-
-        ctx.put("fortunes", fortunes);
-
-        // and now delegate to the engine to render it.
-        engine.render(ctx.data(), "templates/Fortunes.rocker.html", res -> {
-          if (res.succeeded()) {
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "text/html; charset=UTF-8")
-                .end(res.result());
-          } else {
-            ctx.fail(res.cause());
-          }
-        });
-      });
-    }
-
-    public final void updateHandler(final RoutingContext ctx) {
-      final int queries = Helper.getQueries(ctx.request());
-      final World[] worlds = new World[queries];
-
-      new Handler<Integer>() {
-        @Override
-        public void handle(Integer idx) {
-          if (idx == queries) {
-            // stop condition
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encodeToBuffer(worlds));
-
-          } else {
-
-            final Handler<Integer> self = this;
-
-            final int id = randomWorld();
-
-            final JsonObject query = new JsonObject().put("_id", id);
-
-            database.findOne("world", query, FIELDS, findOne -> {
-              if (findOne.failed()) {
-                ctx.fail(findOne.cause());
-                return;
-              }
-
-              final int newRandomNumber = randomWorld();
-
-              database.updateCollection("world", query, new JsonObject().put("$set", new JsonObject().put("randomNumber", newRandomNumber)), update -> {
-                if (update.failed()) {
-                  ctx.fail(update.cause());
-                  return;
-                }
-
-                worlds[idx] = new World(id, newRandomNumber);
-                self.handle(idx + 1);
-              });
-            });
-
-          }
-        }
-      }.handle(0);
-    }
-  }
-
-  /**
-   * PgClient implementation
-   */
-  private final class PgClientBenchmark {
-
-    private static final String UPDATE_WORLD = "UPDATE world SET randomnumber=$1 WHERE id=$2";
-    private static final String SELECT_WORLD = "SELECT id, randomnumber from WORLD where id=$1";
-    private static final String SELECT_FORTUNE = "SELECT id, message from FORTUNE";
-
-    private final PgPool client;
-
-    // In order to use a template we first need to create an engine
-    private final RockerTemplateEngine engine;
-
-    public PgClientBenchmark(Vertx vertx, JsonObject config) {
-      PgConnectOptions options = new PgConnectOptions()
-        .setCachePreparedStatements(true)
-        .setHost(config.getString("host"))
-        .setPort(config.getInteger("port", 5432))
-        .setUser(config.getString("username"))
-        .setPassword(config.getString("password"))
-        .setDatabase(config.getString("database"));
-
-      client = PgPool.pool(vertx, options, new PoolOptions().setMaxSize(4));
-      this.engine = RockerTemplateEngine.create();
-    }
-
-    public final void dbHandler(final RoutingContext ctx) {
-      client
-        .preparedQuery(SELECT_WORLD)
-        .execute(Tuple.of(randomWorld()), res -> {
-        if (res.succeeded()) {
-          final RowIterator<Row> resultSet = res.result().iterator();
-          if (!resultSet.hasNext()) {
-            ctx.response()
-              .setStatusCode(404)
-              .end();
-            return;
-          }
-          final Row row = resultSet.next();
-          ctx.response()
-            .putHeader(HttpHeaders.SERVER, SERVER)
-            .putHeader(HttpHeaders.DATE, date)
-            .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-            .end(Json.encodeToBuffer(new World(row.getInteger(0), row.getInteger(1))));
-        } else {
-          res.cause().printStackTrace();
-          ctx.fail(res.cause());
-        }
-      });
-    }
-
-    public final void queriesHandler(final RoutingContext ctx) {
-
-      final int queries = Helper.getQueries(ctx.request());
-      final World[] worlds = new World[queries];
-      final boolean[] failed = { false };
-      final int[] cnt = { 0 };
-
-      for (int i = 0; i < queries; i++) {
-        client.preparedQuery(SELECT_WORLD).execute(Tuple.of(randomWorld()), res -> {
-          if (!failed[0]) {
-            if (res.failed()) {
-              failed[0] = true;
-              ctx.fail(res.cause());
-              return;
-            }
-
-            // we need a final reference
-            final Row row = res.result().iterator().next();
-            worlds[cnt[0]++] = new World(row.getInteger(0), row.getInteger(1));
-
-            // stop condition
-            if (cnt[0] == queries) {
-              ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encodeToBuffer(worlds));
-            }
-          }
-        });
-      }
-    }
-
-    public final void fortunesHandler(final RoutingContext ctx) {
-
-      client.preparedQuery(SELECT_FORTUNE).execute(ar -> {
-          if (ar.succeeded()) {
-            final RowIterator<Row> resultSet = ar.result().iterator();
-            if (!resultSet.hasNext()) {
-              ctx.fail(404);
-              return;
-            }
-
-            final List<Fortune> fortunes = new ArrayList<>();
-
-            while (resultSet.hasNext()) {
-                final Row row = resultSet.next();
-              fortunes.add(new Fortune(row.getInteger(0), row.getString(1)));
-            }
-
-            fortunes.add(new Fortune(0, "Additional fortune added at request time."));
-            Collections.sort(fortunes);
-
-            ctx.put("fortunes", fortunes);
-
-            // and now delegate to the engine to render it.
-            engine.render(ctx.data(), "templates/Fortunes.rocker.html", res -> {
-              if (res.succeeded()) {
-                ctx.response()
-                        .putHeader(HttpHeaders.SERVER, SERVER)
-                        .putHeader(HttpHeaders.DATE, date)
-                        .putHeader(HttpHeaders.CONTENT_TYPE, "text/html; charset=UTF-8")
-                        .end(res.result());
-              } else {
-                ctx.fail(res.cause());
-              }
-            });
-          } else {
-            ctx.fail(ar.cause());
-          }
-        });
-    }
-
-    public final void updateHandler(final RoutingContext ctx) {
-
-         final int queries = Helper.getQueries(ctx.request());
-         final World[] worlds = new World[queries];
-         final boolean[] failed = { false };
-         final int[] queryCount = { 0 };
-
-         for (int i = 0; i < worlds.length; i++) {
-           int id = randomWorld();
-           client.preparedQuery(SELECT_WORLD).execute(Tuple.of(id), ar2 -> {
-             if (!failed[0]) {
-               if (ar2.failed()) {
-                 failed[0] = true;
-                 ctx.fail(ar2.cause());
-                 return;
-               }
-
-               final Row row = ar2.result().iterator().next();
-               worlds[queryCount[0]++] = new World(row.getInteger(0), randomWorld());
-
-               if (queryCount[0] == worlds.length) {
-                 Arrays.sort(worlds);
-                 List<Tuple> batch = new ArrayList<>();
-
-                 for (World world : worlds) {
-                   batch.add(Tuple.of(world.getRandomNumber(), world.getId()));
-                 }
-
-                 client.preparedQuery(UPDATE_WORLD).executeBatch(batch, ar3 -> {
-                   if (ar3.failed()) {
-                     ctx.fail(ar3.cause());
-                     return;
-                   }
-
-                   ctx.response()
-                     .putHeader(HttpHeaders.SERVER, SERVER)
-                     .putHeader(HttpHeaders.DATE, date)
-                     .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                     .end(Json.encodeToBuffer(worlds));
-                 });
-               }
-             }
-           });
-         }
-    }
-  }
-
-  /**
-   * Implementation using com.github.susom:database library and standard JDBC driver.
-   */
-  private final class DatabaseSqlBenchmark {
-    private final Builder dbBuilder;
-
-    // In order to use a template we first need to create an engine
-    private final RockerTemplateEngine engine;
-
-    DatabaseSqlBenchmark(Vertx vertx, JsonObject jsonConfig) {
-      Config config = ConfigFrom.firstOf().custom(jsonConfig::getString)
-          .rename("username", "database.user")
-          .rename("password", "database.password")
-          .rename("maxPoolSize", "database.pool.size")
-          .value("database.url", "jdbc:postgresql://" + jsonConfig.getString("host") + "/" + jsonConfig.getString("database"))
-          .get();
-      dbBuilder = DatabaseProviderVertx.pooledBuilder(vertx, config);
-      engine = RockerTemplateEngine.create();
-    }
-
-    void dbHandler(final RoutingContext ctx) {
-      dbBuilder.transactAsync(dbs ->
-        dbs.get().toSelect("SELECT id, randomnumber from WORLD where id = ?")
-            .argInteger(randomWorld())
-            .queryFirstOrNull(row -> new World(row.getIntegerOrZero(), row.getIntegerOrZero()))
-      , call -> {
-        if (call.succeeded()) {
-          if (call.result() == null) {
-            ctx.fail(404);
-          } else {
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encode(call.result()));
-          }
-        } else {
-          ctx.fail(call.cause());
-        }
-      });
-    }
-
-    void queriesHandler(final RoutingContext ctx) {
-      int queries = Helper.getQueries(ctx.request());
-
-      dbBuilder.transactAsync(dbs -> {
-        List<World> worlds = new ArrayList<>();
-        for (int i = 1; i <= queries; i++) {
-          dbs.get()
-              .toSelect("SELECT id, randomnumber from WORLD where id = ?")
-              .argInteger(randomWorld())
-              .queryFirstOrNull(row -> worlds.add(new World(row.getIntegerOrZero(), row.getIntegerOrZero())));
-        }
-        return worlds;
-      }, call -> {
-        if (call.succeeded()) {
-          if (call.result() == null) {
-            ctx.fail(404);
-          } else {
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encodeToBuffer(call.result()));
-          }
-        } else {
-          ctx.fail(call.cause());
-        }
-      });
-    }
-
-    final void fortunesHandler(final RoutingContext ctx) {
-      dbBuilder.transactAsync(dbs -> {
-        List<Fortune> fortunes = dbs.get()
-            .toSelect("SELECT id, message from FORTUNE")
-            .queryMany(row -> new Fortune(row.getIntegerOrZero(), row.getStringOrEmpty()));
-        fortunes.add(new Fortune(0, "Additional fortune added at request time."));
-        Collections.sort(fortunes);
-        return fortunes;
-      }, call -> {
-        if (call.succeeded()) {
-          if (call.result() == null || call.result().isEmpty()) {
-            ctx.fail(404);
-          } else {
-            ctx.put("fortunes", call.result());
-            engine.render(ctx.data(), "templates/Fortunes.rocker.html", res -> {
-              if (res.succeeded()) {
-                ctx.response()
-                    .putHeader(HttpHeaders.SERVER, SERVER)
-                    .putHeader(HttpHeaders.DATE, date)
-                    .putHeader(HttpHeaders.CONTENT_TYPE, "text/html; charset=UTF-8")
-                    .end(res.result());
-              } else {
-                ctx.fail(res.cause());
-              }
-            });
-          }
-        } else {
-          ctx.fail(call.cause());
-        }
-      });
-    }
-
-    void updateHandler(final RoutingContext ctx) {
-      int queries = Helper.getQueries(ctx.request());
-
-      dbBuilder.transactAsync(dbs -> {
-        List<World> worlds = new ArrayList<>();
-        // Haven't implemented batching yet on toUpdate(), so hijack toInsert() as work-around
-        SqlInsert batchUpdate = dbs.get().toInsert("UPDATE WORLD SET randomnumber = ? WHERE id = ?");
-        for (int i = 1; i <= queries; i++) {
-          World oldWorld = dbs.get()
-              .toSelect("SELECT id, randomnumber from WORLD where id = ?")
-              .argInteger(randomWorld())
-              .queryFirstOrNull(row -> new World(row.getIntegerOrZero(), row.getIntegerOrZero()));
-          if (oldWorld == null) {
-            return null;
-          }
-          World newWorld = new World(oldWorld.getId(), randomWorld());
-          worlds.add(newWorld);
-          batchUpdate.argInteger(newWorld.getRandomNumber()).argInteger(newWorld.getId()).batch();
-        }
-        batchUpdate.insertBatchUnchecked();
-        return worlds;
-      }, call -> {
-        if (call.succeeded()) {
-          if (call.result() == null) {
-            ctx.fail(404);
-          } else {
-            ctx.response()
-                .putHeader(HttpHeaders.SERVER, SERVER)
-                .putHeader(HttpHeaders.DATE, date)
-                .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .end(Json.encodeToBuffer(call.result()));
-          }
-        } else {
-          ctx.fail(call.cause());
-        }
-      });
-    }
-  }
-
-  private static final String SERVER = "vertx-web";
-  private String date;
+  int port = 8080;
 
   @Override
   public void start() {
     final Router app = Router.router(vertx);
-    // initialize the date header
-    date = DateTimeFormatter.RFC_1123_DATE_TIME.format(ZonedDateTime.now());
-    // refresh the value as a periodic task
-    vertx.setPeriodic(1000, handler -> date = DateTimeFormatter.RFC_1123_DATE_TIME.format(ZonedDateTime.now()));
 
     final MongoDBBenchmark mongoDBBenchmark = new MongoDBBenchmark(vertx, config());
     final PgClientBenchmark pgClientBenchmark= new PgClientBenchmark(vertx, config());
     final DatabaseSqlBenchmark databaseSqlBenchmark = new DatabaseSqlBenchmark(vertx, config());
+    final JsonBenchmark jsonBenchmark = new JsonBenchmark(vertx);
+    final PlaintextBenchmark plaintextBenchmark = new PlaintextBenchmark(vertx);
 
     /*
      * This test exercises the framework fundamentals including keep-alive support, request routing, request header
      * parsing, object instantiation, JSON serialization, response header generation, and request count throughput.
      */
-    app.get("/json").handler(ctx -> {
-      ctx.response()
-          .putHeader(HttpHeaders.SERVER, SERVER)
-          .putHeader(HttpHeaders.DATE, date)
-          .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-          .end(Json.encodeToBuffer(new Message("Hello, World!")));
-    });
+    app.get("/json").handler(jsonBenchmark::json);
 
     /*
      * This test exercises the framework's object-relational mapper (ORM), random number generator, database driver,
@@ -560,19 +64,17 @@ public class App extends AbstractVerticle {
      * still small, meaning good performance is still necessary in order to saturate the gigabit Ethernet of the test
      * environment.
      */
-    app.get("/plaintext").handler(ctx -> {
-      ctx.response()
-          .putHeader(HttpHeaders.SERVER, SERVER)
-          .putHeader(HttpHeaders.DATE, date)
-          .putHeader(HttpHeaders.CONTENT_TYPE, "text/plain")
-          .end("Hello, World!");
-    });
+    app.get("/plaintext").handler(plaintextBenchmark::plaintext);
 
-    vertx.createHttpServer().requestHandler(app).listen(8080, listen -> {
+    vertx.createHttpServer().requestHandler(app).listen(port, listen -> {
       if (listen.failed()) {
         listen.cause().printStackTrace();
         System.exit(1);
+      }else {
+        System.out.println("Vert.x run on port:"+port);
       }
     });
+
   }
+
 }
