+++ frameworks/Java/vertx-web/src/main/java/io/vertx/benchmark/benchmarks/Benchmark.java
@@ -1,0 +1,15 @@
+package io.vertx.benchmark.benchmarks;
+
+import io.vertx.ext.web.RoutingContext;
+
+public interface Benchmark {
+
+    public void dbHandler(final RoutingContext ctx);
+
+    public void queriesHandler(final RoutingContext ctx);
+
+    public void fortunesHandler(final RoutingContext ctx);
+
+    public void updateHandler(final RoutingContext ctx);
+
+}
